requirements.txt
Pillow
imagehash
numpy